+++ docs/changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 ---------
-Unreleased
-^^^^^^^^^^
+2.0.7
+^^^^^
 - Implemented skipping duplicate special characters for SCCReader
 - Added support for beautifulsoup 4.10 and lxml 4.8
 - Added pytest and pytest-lazy-fixture as development dependencies